Added 'soleValue' method for lazy collections.
